rearrange files per review comment
